Run covr monthly.
Note: This will throw an error until we get covrpage working properly on this repo, but it feels like the right place to put a monthly check.

The monthly check will confirm that the API is working as expected for this package, and will regenerate the test messages on the test slack.